Cleanup Pixel Launcher script

+++ pixellauncher-wallpaper.sh
@@ -15,42 +15,41 @@ echo "" >> build.log
 echo "Updating "$INDIR" on $DATE for lollipop, marshmallow, and nougat" >> build.log
 echo "Google Pixel Launcher add-on for 5.0.2+ (arm/arm64) (includes Gooogle Wallpapers)" >> build.log
 
-DIR1=$(dirname "${FILEPATH1}")
-DIR2=$(dirname "${FILEPATH2}")
-FILE1=${FILEPATH1##*/}
-FILE2=${FILEPATH2##*/}
-NOEXT1=${FILE1%\.*}
-NOEXT2=${FILE2%\.*}
-
-cd "$DIR1"
-
-if [ ! $FILE1 == "" ]; then
-  rm "$FILENAME1"
-  mv "$FILE1" "$FILENAME1"
+if [ ! "$FILEPATH1" == "" ]; then
+  DIR1=$(dirname "${FILEPATH1}")
+  FILE1=${FILEPATH1##*/}
+  NOEXT1=${FILE1%\.*}
+  cd "$DIR1"
+  if [ ! $FILE1 == "" ]; then
+    rm "$FILENAME1"
+    mv "$FILE1" "$FILENAME1"
+    VERSION1=$(echo "$FILE1" | cut -d "_" -f 2)
+    APIVER1=$(echo "$FILE1" | cut -d "_" -f 3)
+    cd "$BASEDIR"
+    echo "" >> build.log
+    echo "Updating Google Wallpapers" >> build.log
+    echo "Version: $VERSION1" >> build.log
+    echo "API: $APIVER1" >> build.log
+  fi
 fi
 
-cd "$BASEDIR"
-cd "$DIR2"
-
-if [ ! $FILE2 == "" ]; then
-  rm "$FILENAME2"
-  mv "$FILE2" "$FILENAME2"
+if [ ! "$FILEPATH2" == "" ]; then
+  DIR2=$(dirname "${FILEPATH2}")
+  FILE2=${FILEPATH2##*/}
+  NOEXT2=${FILE2%\.*}
+  cd "$BASEDIR"
+  cd "$DIR2"
+  if [ ! $FILE2 == "" ]; then
+    rm "$FILENAME2"
+    mv "$FILE2" "$FILENAME2"
+    VERSION2=$(echo "$FILE2" | cut -d "_" -f 2)
+    APIVER2=$(echo "$FILE2" | cut -d "_" -f 3)
+    cd "$BASEDIR"
+    echo "" >> build.log
+    echo "Updating Google Pixel Launcher" >> build.log
+    echo "Version: $VERSION2" >> build.log
+    echo "API: $APIVER2" >> build.log
+  fi
 fi
-
-VERSION1=$(echo "$FILE1" | cut -d "_" -f 2)
-APIVER1=$(echo "$FILE1" | cut -d "_" -f 3)
-VERSION2=$(echo "$FILE2" | cut -d "_" -f 2)
-APIVER2=$(echo "$FILE2" | cut -d "_" -f 3)
-
-cd "$BASEDIR"
-
-echo "Google Wallpapers" >> build.log
-echo "Version: $VERSION1" >> build.log
-echo "API: $APIVER1" >> build.log
-echo "" >> build.log
-echo "Google Pixel Launcher" >> build.log
-echo "Version: $VERSION2" >> build.log
-echo "API: $APIVER2" >> build.log
-echo "" >> build.log
 
 ./makezipsign.sh "$INDIR" "$NAME" "$VER"
